[refactor] Move local to global component
